+++ src/components/LogInPage/LogInPageLeft/index.js
@@ -1,0 +1,22 @@
+import React from 'react'
+import * as S from './style';
+
+const LogInPageLeft = () => {
+    return (
+        <S.LogInPageLeft>
+        <S.LeftMainTextWrapper>
+          <S.MainTitle color="white" fontSize="big">
+            당신의 개발 동료를 찾아줄
+          </S.MainTitle>
+          <S.MainTitle color="white" fontSize="big">
+            최고의 온라인 커뮤니티
+          </S.MainTitle>
+          <S.MainTitle margin="top" color="white" fontSize="middle">
+            Rubminds에서 당신의 팀원을 찾으세요!
+          </S.MainTitle>
+        </S.LeftMainTextWrapper>
+      </S.LogInPageLeft>
+    )
+}
+
+export default LogInPageLeft;
