--- src/components/Writting/InputNumber.js
@@ -1,3 +1,0 @@
-import InputNumber from "rc-input-number";
-
-export default () => <InputNumber defaultValue={10}></InputNumber>
